Add files via upload
here is the task of fecting the data from the url and show them to in a app. 




here is a task file but it not run i don't know why i think i mistake in code can you please check it  what i mistake. and please give tell me what i mistake it really help me to learn new technology thankyou..